Fix/manifest

+++ sonia_navigation_states/src/sonia_navigation_states/search_zigzag.py
@@ -55,7 +55,7 @@ class search_zigzag(EventState):
         self.side =side
 
         # Compute trajectory parameters
-        self.fullStep = math.floor(self.boxX/self.stroke)
+        self.fullStep = int(math.floor(self.boxX/self.stroke))
         self.residue = self.boxX % self.stroke
